Fix RawSwerve bug with AutoXChassis
RawSwerve took in 1-byte values and added them up in order to
calculate speed. However, if the sum of x, y, and turn was ever greater
to 127, then it would overflow and the chassis would act weird. To fix
this, simply use 2-byte values when adding.

+++ src/inu/auto/chassis/AutoXChassis.cpp
@@ -4,6 +4,7 @@
 #include "inu/motor/background/PIDInertialMotor.h"
 #include "inu/InuException.hpp"
 #include "inu/background/BackgroundSystem.h"
+#include <cstdint>
 
 using namespace inu;
 
@@ -83,10 +84,10 @@ void AutoXChassis::RawSwerve(std::int8_t forward, std::int8_t right, std::int8_t
 	right = std::clamp<std::int8_t>(right, -127, 127);
 	turn = std::clamp<std::int8_t>(turn, -127, 127);
 
-	auto tl = std::clamp<std::int8_t>(forward + turn + right, -127, 127);
-	auto bl = std::clamp<std::int8_t>(forward + turn - right, -127, 127);
-	auto tr = std::clamp<std::int8_t>(-forward + turn + right, -127, 127);
-	auto br = std::clamp<std::int8_t>(-forward + turn - right, -127, 127);
+	auto tl = std::clamp<std::int16_t>(forward + turn + right, -127, 127);
+	auto bl = std::clamp<std::int16_t>(forward + turn - right, -127, 127);
+	auto tr = std::clamp<std::int16_t>(-forward + turn + right, -127, 127);
+	auto br = std::clamp<std::int16_t>(-forward + turn - right, -127, 127);
 
 	topleft->Move(tl);
 	topright->Move(tr);
